remove unused and rename files
