--- .eslintrc.js
@@ -1,36 +1,0 @@
-module.exports = {
-  extends: ['airbnb-typescript'],
-  overrides: [
-    {
-      files: [
-        'packages/components/src/**/**/*.ts',
-        'packages/components/src/**/**/*.tsx',
-        'packages/hooks/src/**/**/*.ts',
-        'packages/hooks/src/**/**/*.tsx',
-        'packages/sdk/src/**/**/*.ts',
-        'packages/sdk/src/**/**/*.tsx',
-        'packages/utils/src/**/**/*.ts',
-        'packages/utils/src/**/**/*.tsx',
-      ],
-      rules: {
-        'import/no-extraneous-dependencies': 'off',
-        'import/prefer-default-export': 'off',
-        'implicit-arrow-linebreak': 'off',
-        'function-paren-newline': 'off',
-        'max-len': 'off',
-        'react/prop-types': 'off',
-        'react/jsx-props-no-spreading': 'off',
-        // prettier takes care
-        'object-curly-newline': 'off',
-        'arrow-parens': 'off',
-      },
-    },
-  ],
-  env: {
-    jest: true,
-    browser: true,
-  },
-  parserOptions: {
-    project: './tsconfig.json',
-  },
-};
