+++ docs/.eslintrc
@@ -1,6 +1,4 @@
 {
-  // comment out because it gave a warning
-  // "extends": ["@sampotts/eslint-config/react-ts"],
   "env": {
     "browser": true
   },
