+++ docs/pages/index.mdx
@@ -1,10 +1,6 @@
 import SEO from '../components/SEO';
-import snippets from '../components/snippets';
 
-<SEO
-  title="Getting Started"
-  description="TBD"
-/>
+<SEO title="Getting Started" description="TBD" />
 
 # Getting Started
