+++ packages/components/src/Checkbox/index.tsx
@@ -4,7 +4,7 @@ import { useId } from '@reach/auto-id';
 import React from 'react';
 import tw from 'twin.macro';
 import useInputStyle, { UseInputStyleProps } from '../hooks/use-input-styles';
-import { __DEV__ } from '../utils/assersion';
+import { __DEV__ } from '../utils/assertion';
 import Label from '../Label';
 import { CheckboxProps } from './types';
