+++ packages/components/src/Radio/index.tsx
@@ -5,7 +5,7 @@ import { useId } from '@reach/auto-id';
 import React from 'react';
 import tw from 'twin.macro';
 import useInputStyle, { UseInputStyleProps } from '../hooks/use-input-styles';
-import { __DEV__ } from '../utils/assersion';
+import { __DEV__ } from '../utils/assertion';
 import Label from '../Label';
 import { RadioProps } from './types';
