+++ packages/components/src/Tabs/TabPanels/index.tsx
@@ -1,6 +1,6 @@
 import React, { cloneElement } from 'react';
 
-import { __DEV__ } from '../../utils/assersion';
+import { __DEV__ } from '../../utils/assertion';
 import { cleanChildren } from '../../utils/react-helpers';
 import Box from '../../Box';
 import { useTabContext } from '../context';
