+++ packages/components/src/utils/object.ts
@@ -1,5 +1,5 @@
 /* eslint-disable no-param-reassign */
-import { isObject } from './assersion';
+import { isObject } from './assertion';
 
 interface MergeObject {
   [k: string]: any;
